apps/eth: allow larger chain IDs in the RLP encoding
Integers larger than 0x7f have a length prefix in RLP encoding. Until
now, we only had chain IDs smaller than 0x7f, so we could directly
encode it without a length prefix. We want to add more networks with
larger chain IDs, so we have to RLP encode these as well.

+++ src/rust/bitbox02-rust/src/hww/api/ethereum/params.rs
@@ -21,7 +21,7 @@ pub struct Params {
     pub coin: EthCoin,
     pub bip44_coin: u32,
     /// https://github.com/ethereum/EIPs/blob/master/EIPS/eip-155.md#list-of-chain-ids
-    pub chain_id: u8,
+    pub chain_id: u64,
     pub name: &'static str,
     pub unit: &'static str,
 }
